+++ src/AutoRest.CSharp/Mgmt/Decorator/SchemaExtensions.cs
@@ -16,7 +16,7 @@ namespace AutoRest.CSharp.Mgmt.Decorator
         /// </summary>
         /// <param name="schema"></param>
         /// <returns></returns>
-        private static IEnumerable<Property> GetAllProperties(this ObjectSchema schema)
+        internal static IEnumerable<Property> GetAllProperties(this ObjectSchema schema)
         {
             return schema.Parents!.All.OfType<ObjectSchema>().SelectMany(parentSchema => parentSchema.Properties).Concat(schema.Properties);
         }
